Change unit test to not rely on implementation-level details
The purpose of this test was to assert that conf_accept() consumes
the input. The first method was to check the internal state's buffer,
but that's a bad practice for unit testing. This test is better, I think,
because it tries to read more and ensure that it cannot. But even still,
I think there's a better way to define this test. We'll see.