Consider whitespace significant; only skip comments
I'm not 100% sure about this decision because it puts more onus
on the library end-user to be cognizant of her whitespace and
parse that.

In order to support newlines surrounded by newlines, it became
necessary to do this:

    while (conf_accept("\n"));

Hopefully, by having ght end-user manage their own whitespace, it
makes libconf more flexible and faster(?).